AIML-1394 Move ApplicationLicenseDiscriminator to tool.base
Move ApplicationLicenseDiscriminator from tool.application to tool.base
to eliminate the cross-domain dependency from tool.coverage. Introduce
ContrastAccessDeniedException in the client package to replace the SDK
UnauthorizedException parameter, fixing the SDK containment violation.
