+++ contrast-mcp-core/src/main/java/com/contrast/labs/ai/mcp/contrast/client/ContrastAccessDeniedException.java
@@ -1,0 +1,40 @@
+/*
+ * Copyright 2026 Contrast Security
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.contrast.labs.ai.mcp.contrast.client;
+
+import com.contrastsecurity.exceptions.UnauthorizedException;
+import lombok.Getter;
+
+/**
+ * SDK-independent representation of an HTTP 403 from the Contrast API. Wraps the SDK {@link
+ * UnauthorizedException} so tool-layer code can handle access denials without importing the SDK
+ * directly.
+ */
+@Getter
+public class ContrastAccessDeniedException extends RuntimeException {
+
+  private final int httpCode;
+
+  private ContrastAccessDeniedException(String message, int httpCode, Throwable cause) {
+    super(message, cause);
+    this.httpCode = httpCode;
+  }
+
+  public static ContrastAccessDeniedException from(UnauthorizedException sdkException) {
+    return new ContrastAccessDeniedException(
+        sdkException.getMessage(), sdkException.getCode(), sdkException);
+  }
+}
