+++ contrast-mcp-core/src/test/resources/archunit-store/99efc868-3313-4f9a-948d-23c14f9ac1e1
@@ -27,7 +27,6 @@ Method <com.contrast.labs.ai.mcp.contrast.result.ScanProject.from(com.contrastse
 Method <com.contrast.labs.ai.mcp.contrast.result.ScanProject.from(com.contrastsecurity.sdk.scan.Project)> has parameter of type <com.contrastsecurity.sdk.scan.Project> in (ScanProject.java:0)
 Method <com.contrast.labs.ai.mcp.contrast.result.VulnLight.sessionMetadata()> has generic return type <java.util.List<com.contrastsecurity.models.SessionMetadata>> with type argument depending on <com.contrastsecurity.models.SessionMetadata> in (VulnLight.java:0)
 Method <com.contrast.labs.ai.mcp.contrast.result.Vulnerability.sessionMetadata()> has generic return type <java.util.List<com.contrastsecurity.models.SessionMetadata>> with type argument depending on <com.contrastsecurity.models.SessionMetadata> in (Vulnerability.java:0)
-Method <com.contrast.labs.ai.mcp.contrast.tool.application.ApplicationLicenseDiscriminator.discriminate(java.lang.String, com.contrastsecurity.exceptions.UnauthorizedException)> has parameter of type <com.contrastsecurity.exceptions.UnauthorizedException> in (ApplicationLicenseDiscriminator.java:0)
 Method <com.contrast.labs.ai.mcp.contrast.tool.application.GetSessionMetadataTool.doExecute(com.contrast.labs.ai.mcp.contrast.tool.application.params.GetSessionMetadataParams, com.contrast.labs.ai.mcp.contrast.tool.base.NoticeCollector)> has return type <com.contrastsecurity.models.MetadataFilterResponse> in (GetSessionMetadataTool.java:0)
 Method <com.contrast.labs.ai.mcp.contrast.tool.application.GetSessionMetadataTool.getSessionMetadata(java.lang.String)> has generic return type <com.contrast.labs.ai.mcp.contrast.tool.base.SingleToolResponse<com.contrastsecurity.models.MetadataFilterResponse>> with type argument depending on <com.contrastsecurity.models.MetadataFilterResponse> in (GetSessionMetadataTool.java:0)
 Method <com.contrast.labs.ai.mcp.contrast.tool.application.GetSessionMetadataTool.getSessionMetadata(java.lang.String, org.springframework.ai.chat.model.ToolContext)> has generic return type <com.contrast.labs.ai.mcp.contrast.tool.base.SingleToolResponse<com.contrastsecurity.models.MetadataFilterResponse>> with type argument depending on <com.contrastsecurity.models.MetadataFilterResponse> in (GetSessionMetadataTool.java:0)
