--- contrast-mcp-core/src/test/resources/archunit-store/b141356a-0184-40ad-b0c0-624576f073da
@@ -1,11 +1,0 @@
-Constructor <com.contrast.labs.ai.mcp.contrast.tool.coverage.GetRouteCoverageTool.<init>(com.contrast.labs.ai.mcp.contrast.client.ContrastApiClient, com.contrast.labs.ai.mcp.contrast.tool.coverage.RouteMapper, com.contrast.labs.ai.mcp.contrast.tool.application.ApplicationLicenseDiscriminator)> has parameter of type <com.contrast.labs.ai.mcp.contrast.tool.application.ApplicationLicenseDiscriminator> in (GetRouteCoverageTool.java:0)
-Field <com.contrast.labs.ai.mcp.contrast.tool.coverage.GetRouteCoverageTool.applicationLicenseDiscriminator> has type <com.contrast.labs.ai.mcp.contrast.tool.application.ApplicationLicenseDiscriminator> in (GetRouteCoverageTool.java:0)
-Method <com.contrast.labs.ai.mcp.contrast.tool.coverage.GetRouteCoverageTool.fetchRouteCoverage(java.lang.String, com.contrast.labs.ai.mcp.contrast.sdkextension.data.routecoverage.RouteCoverageBySessionIDAndMetadataRequestExtended)> calls method <com.contrast.labs.ai.mcp.contrast.tool.application.ApplicationLicenseDiscriminator$ApplicationState.ordinal()> in (GetRouteCoverageTool.java:201)
-Method <com.contrast.labs.ai.mcp.contrast.tool.coverage.GetRouteCoverageTool.fetchRouteCoverage(java.lang.String, com.contrast.labs.ai.mcp.contrast.sdkextension.data.routecoverage.RouteCoverageBySessionIDAndMetadataRequestExtended)> calls method <com.contrast.labs.ai.mcp.contrast.tool.application.ApplicationLicenseDiscriminator.discriminate(java.lang.String, com.contrastsecurity.exceptions.UnauthorizedException)> in (GetRouteCoverageTool.java:199)
-Static Initializer <com.contrast.labs.ai.mcp.contrast.tool.coverage.GetRouteCoverageTool$1.<clinit>()> calls method <com.contrast.labs.ai.mcp.contrast.tool.application.ApplicationLicenseDiscriminator$ApplicationState.ordinal()> in (GetRouteCoverageTool.java:201)
-Static Initializer <com.contrast.labs.ai.mcp.contrast.tool.coverage.GetRouteCoverageTool$1.<clinit>()> calls method <com.contrast.labs.ai.mcp.contrast.tool.application.ApplicationLicenseDiscriminator$ApplicationState.ordinal()> in (GetRouteCoverageTool.java:201)
-Static Initializer <com.contrast.labs.ai.mcp.contrast.tool.coverage.GetRouteCoverageTool$1.<clinit>()> calls method <com.contrast.labs.ai.mcp.contrast.tool.application.ApplicationLicenseDiscriminator$ApplicationState.ordinal()> in (GetRouteCoverageTool.java:201)
-Static Initializer <com.contrast.labs.ai.mcp.contrast.tool.coverage.GetRouteCoverageTool$1.<clinit>()> calls method <com.contrast.labs.ai.mcp.contrast.tool.application.ApplicationLicenseDiscriminator$ApplicationState.values()> in (GetRouteCoverageTool.java:201)
-Static Initializer <com.contrast.labs.ai.mcp.contrast.tool.coverage.GetRouteCoverageTool$1.<clinit>()> gets field <com.contrast.labs.ai.mcp.contrast.tool.application.ApplicationLicenseDiscriminator$ApplicationState.ARCHIVED> in (GetRouteCoverageTool.java:201)
-Static Initializer <com.contrast.labs.ai.mcp.contrast.tool.coverage.GetRouteCoverageTool$1.<clinit>()> gets field <com.contrast.labs.ai.mcp.contrast.tool.application.ApplicationLicenseDiscriminator$ApplicationState.LICENSED> in (GetRouteCoverageTool.java:201)
-Static Initializer <com.contrast.labs.ai.mcp.contrast.tool.coverage.GetRouteCoverageTool$1.<clinit>()> gets field <com.contrast.labs.ai.mcp.contrast.tool.application.ApplicationLicenseDiscriminator$ApplicationState.UNLICENSED> in (GetRouteCoverageTool.java:201)
